requirements.txt
asgiref==3.8.1
Django==5.1.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.45
PyJWT==2.9.0
sqlparse==0.5.1
